% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{bom}
\alias{bom}
\title{Fit Bayesian Ordinary Differential Equations (ODE) Multilevel Models}
\usage{
bom(model, effect = NULL, prior, data, family, sample_prior = c("no",
  "yes", "only"), chains = 1, iter = 2000, warmup = floor(iter/2),
  thin = 1, cores = getOption("mc.cores", 1L), control = NULL,
  seed = NA, ...)
}
\description{
Fit Bayesian Ordinary Differential Equations (ODE) Multilevel Models
}
