% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{fitode}
\alias{fitode}
\title{fit ode}
\usage{
fitode(model, data, init, tcol = "times", t0, method = c("optimizing",
  "sampling"), link, fixed, ...)
}
\arguments{
\item{model}{ode model}

\item{data}{data frame with time column and observation column}

\item{init}{named vector of initial parameter values}

\item{tcol}{time column}

\item{method}{fit method}

\item{link}{named vector or list of link functions for ode/log-likelihood parameters}

\item{...}{rstan arguments}

\item{debug}{print debugging output?}
}
\description{
fit ode
}
\seealso{
\code{\link{mle2}}
}
