% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{initialize,model.ode-method}
\alias{initialize,model.ode-method}
\title{the initializer for model.ode}
\usage{
\S4method{initialize}{model.ode}(.Object, name, model, observation,
  initial, par)
}
\arguments{
\item{.Object}{object}
}
\description{
the initializer for model.ode
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{model}}{ode model}

\item{\code{observation}}{observation model}

\item{\code{initial}}{initial values}

\item{\code{par}}{parameters}
}}

\examples{
SI_model <- new("model.ode",
    name = "SI",
    model = list(
        S ~ - beta*S*I/N,
        I ~ beta*S*I/N - gamma*I
    ),
    observation = list(
        susceptible ~ normal(mean=S, sd=sigma1),
        infected ~ normal(mean=I, sd=sigma2)
    ),
    initial = list(
        S ~ N * (1 - i0),
        I ~ N * i0
    ),
    par= c("beta", "gamma", "N", "i0", "sigma1", "sigma2")
)
}
