% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R, R/fitode.R
\docType{class}
\name{fitode}
\alias{fitode}
\title{Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation}
\usage{
fitode(model, data, init, tcol = "times", t0, method = c("optimizing",
  "sampling"), link, fixed, ...)
}
\arguments{
\item{model}{ode model}

\item{data}{data frame with time column and observation column}

\item{init}{named vector of initial parameter values}

\item{tcol}{time column}

\item{method}{optimization method}

\item{link}{named vector or list of link functions for ode/log-likelihood parameters}

\item{...}{mle2 arguments}

\item{optimizer}{optimizer}

\item{control}{see optim}

\item{solver.opts}{options for ode integration. See ode}

\item{skip.hessian}{skip hessian calculation}

\item{use.ginv}{use generalized inverse (\code{\link{ginv}}) to compute approximate vcov}

\item{debug}{print debugging output?}
}
\description{
Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation

fit ode
}
\seealso{
\code{\link{mle2-class}}

\code{\link{mle2}}
}
